backend/storage/models/project_1/requirements.txt
torch>=2.2
pandas
pyyaml
